Moved Dispatchers back to action-creators. Moved tests out of duplicate folders.
