+++ app/containers/RoutesContainer.jsx
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import Routes from '../routes';
-import { getProducts, getProductById } from '../thunks-dispatchers/products-dispatchers';
+import { getProducts, getProductById } from '../action-creators/products';
 // import { retrieveLoggedInUser } from '../action-creators/auth';
 
 const mapState = null;
